v1.2.1
There was an issue with the correctHyphenatedTerm function when using the 'apa' style. The function incorrectly converted hyphenated words to lowercase even when they were not short conjunctions, prepositions, articles, or words that should not be capitalized. This caused words like "server-side" to be incorrectly converted to "server-side" instead of "Server-Side".

The issue was fixed by adding an additional else if statement to the apa case in the switch statement in correctHyphenatedTerm. The new else if statement checks if the hyphenated word is not a short word and should be capitalized. If so, it capitalizes the first letter of the word.